cloud_functions/agent_engine_stream/requirements.txt
# Cloud Functions用 Agent Engineストリーミング機能の依存関係
# 2025年対応 Vertex AI SDK とCloud Functions必須パッケージ

# Google Agent Development Kit (ADK)
google-adk~=1.5.0

# Vertex AI 生成AI SDK（Agent Engine API対応）
google-cloud-aiplatform>=1.91.0

# Cloud Functions Framework（第2世代対応・ストリーミング機能）
functions-framework>=3.7.0

# HTTP リクエスト処理（ストリーミング対応）
requests>=2.31.0
httpx>=0.27.0

# JSON・データ処理
pydantic>=2.0.0

# Google Cloud 認証・ロギング
google-auth>=2.23.0
google-cloud-logging>=3.11.4

# Firebase Admin SDK（認証）
firebase-admin>=6.9.0

# 環境変数・設定管理
python-dotenv>=1.0.0

# Flask（functions-framework依存）
flask>=3.0.0